requirements.txt
Flask
python-dotenv
gunicorn
python-telegram-bot==13.15